+++ toolbox/diver/lines-count.sh
@@ -7,21 +7,14 @@
 #
 # Author: Pierre-Yves LAPERSONNE <pierreyves(dot)lapersonne(at)orange(dot)com> et al.
 
-# Version.............: 1.0.0
+# Version.............: 2.0.0
 # Since...............: 18/07/2023
-# Description.........: Counts liens of cound for a target directory
-#
-# Usage: bash lines-count.sh --target TARGET
-#
-# Exit codes:
-#       0 - normal exit
-#       1 - problem with given parameters
-#
+# Description.........: Counts lines of code for a target directory or Git repo URL
 
 set -euo pipefail # set -euxo pipefail
 
-VERSION="1.0.0"
-SCRIPT_NAME="lines-count.sh"
+VERSION="2.0.0"
+SCRIPT_NAME=$(basename "$0")
 
 # -------------
 # Configuration
@@ -29,6 +22,7 @@ SCRIPT_NAME="lines-count.sh"
 
 NORMAL_EXIT_CODE=0
 BAD_ARGUMENTS_EXIT_CODE=1
+BAD_PRECONDITIONS_EXIT_CODE=2
 
 # Folder fo generated files
 TEMP_FOLDER="./data"
@@ -38,6 +32,9 @@ GENERATED_FILES_PREFIX="$$-lines-count"
 
 # Report file with metrics
 REPORT_METRIC_FILE="$TEMP_FOLDER/$GENERATED_FILES_PREFIX-report.txt"
+
+# Name of clone repository if done
+REPO_CLONE_NAME="$TEMP_FOLDER/repo-to-scan"
 
 # ---------
 # Functions
@@ -50,8 +47,13 @@ DisplayUsages(){
     echo "$SCRIPT_NAME - Version $VERSION"
     echo "***********************************************"
     echo "USAGE:"
-    echo "bash $SCRIPT_NAME --target TARGET"
-    echo -e "\t --target.........: TARGET must point to a directory to scan for lines and count them (recursive)"
+    echo "bash $SCRIPT_NAME [--folder TARGET | --url URL_OF_GIT_REPO]"
+    echo -e "\t --folder.........: TARGET must point to a directory to scan for lines and count them (recursive)"
+    echo -e "\t --url............: Clone repo at URL_OF_GIT_REPO and scan it"
+    echo "EXIT CODES:"
+    echo -e "\t$NORMAL_EXIT_CODE: normal exit"
+    echo -e "\t$BAD_ARGUMENTS_EXIT_CODE: problem with given parameters"
+    echo -e "\t$BAD_PRECONDITIONS_EXIT_CODE: preconditions issues"
 }
 
 # \fn NormalExit
@@ -69,14 +71,17 @@ BadArgumentsExit(){
 # \fn BadPreconditionsExit
 # \brief Exits with BAD_PRECONDITION_EXIT_CODE code
 BadPreconditionsExit() {
-    exit $BAD_PRECONDITION_EXIT_CODE
+    exit $BAD_PRECONDITIONS_EXIT_CODE
 }
 
 # \fn CleanFiles
 # \brief If existing removes the work files
 CleanFiles() {
-    if [ -f $REPORT_METRIC_FILE ]; then
-        rm $REPORT_METRIC_FILE
+    if [ -f "$REPORT_METRIC_FILE" ]; then
+        rm "$REPORT_METRIC_FILE"
+    fi
+    if [ -d "$REPO_CLONE_NAME" ]; then
+        rm -rf "$REPO_CLONE_NAME"
     fi
 }
 
@@ -98,9 +103,23 @@ if [ "$#" -ne 2 ]; then
 fi
 
 # Get target
-if [ "$1" = "--target"  ]; then
+if [ "$1" = "--folder"  ]; then
     if [ "$2" ]; then
 	    directory_to_scan=$2
+        repo_to_clone_url=""
+        # Check if target is directory
+        if [ ! -d "$directory_to_scan" ]; then
+            echo "💥 Error: Target is not a directory ($directory_to_scan)."
+            BadArgumentsExit
+        fi
+    else
+        DisplayUsages
+        BadArgumentsExit
+    fi
+elif [ "$1" = "--url"  ]; then
+    if [ "$2" ]; then
+        directory_to_scan=""
+	    repo_to_clone_url=$2
     else
         DisplayUsages
         BadArgumentsExit
@@ -123,26 +142,46 @@ echo "**************************************************"
 echo "$SCRIPT_NAME - Version $VERSION"
 echo "**************************************************"
 
-echo -e "\n"
-
-echo "📋 Directory to scan is to analyse is $directory_to_scan"
 echo "📋 Prepare logs"
 PrepareFiles
 
-echo -e "\n"
+# ---------------------------------
+# Step 1 - If URL of repo, clone it
+# ---------------------------------
+
+if [ ! -z "$repo_to_clone_url" ]; then
+    echo "📋 Repository to clone is '$repo_to_clone_url'"
+    git clone "$repo_to_clone_url" "$REPO_CLONE_NAME"
+    if [ "$(ls -A $REPO_CLONE_NAME)" ]; then
+        echo "👌 Repository after cloning is not empty, good!"
+        directory_to_scan="$REPO_CLONE_NAME"
+    else
+        echo "💥 Error: After cloning repository is empty, cannot compute metric"
+        BadPreconditionsExit
+    fi
+else
+    echo "📋 No repository to clone, check suplied folder"
+fi
+
+echo "📋 Directory to scan is $directory_to_scan"
 
 # ------------------------------------------------
-# Step 1 - Count lines of target and store results
+# Step 2 - Count lines of target and store results
 # ------------------------------------------------
 
 echo "🍥 Counting lines and store in $directory_to_scan"
+
+if [ ! "$(ls -A $directory_to_scan)" ]; then
+    echo "💥 Error: Directory to scan is empty, cannot compute metric"
+    BadPreconditionsExit
+fi
 
 cloc "$directory_to_scan" > "$REPORT_METRIC_FILE"
 
 echo -e "👌 Counting of lines is done\n"
 
 # ------------------------------------------------------------
-# Step 2 - Metrics (words and files counts, hits, duration...)
+# Step 3 - Metrics (words and files counts, hits, duration...)
 # ------------------------------------------------------------
 
 script_duration=$SECONDS